% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesDCirc.R
\name{BayesDCirc}
\alias{BayesDCirc}
\title{gibbs sampling main function}
\usage{
BayesDCirc(
  Y1,
  Y2,
  t1,
  t2,
  n.iter,
  hyperparameters,
  showIteration,
  priorInformation = NULL,
  priorFDR = NULL,
  Period = 24
)
}
\arguments{
\item{Y1}{expression values of group 1. It needs unique row names.}

\item{Y2}{expression values of group 2. It needs unique row names.}

\item{t1}{phenotype data of group 1}

\item{t2}{phenotype data of group 2}

\item{n.iter}{number of iteration. Default is 10000.}

\item{hyperparameters}{hyperparameters we need}

\item{showIteration}{logic value, if need show iteration when processing gibbs sampling}

\item{priorInformation}{a data frame with three columns. The first is gene's name, the second is the indicator for differential circadian gene with pattern difference, and the third is the indicator for differential circadian gene with MESOR difference. Default is NULL.}

\item{priorFDR}{The FDR cut off that prior information was used to claim differential circadian genes. Default is NULL.}

\item{Period}{numeric value. The time period for which the data needs to be estimated. Default is 24.}
}
\value{
a list that contains posterior values
}
\description{
Detecting differential circadian pattern using gibbs sampling
}
\details{
We want to use gibbs sampling to sample from posterior distributions for our parameter
}
\author{
Yutao Zhang
}
