% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFDR.R
\name{BayesianFDR}
\alias{BayesianFDR}
\title{compute the Bayesian FDR}
\usage{
BayesianFDR(output_BayesDCirc, indicator)
}
\arguments{
\item{output_BayesDCirc}{a list containing posterior values, typically obtained from \code{BayesDCirc}.}

\item{indicator}{a character string, either "I" or "L", indicating which specific indicator to check.}
}
\value{
a data frame containing FDR of all gene candidates.
}
\description{
BayesianFDR Function
}
\details{
We want to compute FDR using this function
}
\examples{
BayesianFDR(output, indicator = "I")
}
\author{
Yutao Zhang
}
