% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hyperparameters.R
\name{get_hyperparameters}
\alias{get_hyperparameters}
\title{gibbs sampling main function}
\usage{
get_hyperparameters(
  Y1,
  Y2,
  t1,
  t2,
  use_default = T,
  personalized_para = NULL,
  Period = 24
)
}
\arguments{
\item{Y1}{expression values of group 1}

\item{Y2}{expression values of group 2}

\item{t1}{phenotype data of group 1}

\item{t2}{phenotype data of group 2}

\item{use_default}{logic value, if the default/non-informative values for the prior are used. It does not implement a spike-and-slab prior. Default is TRUE.}

\item{personalized_para}{numeric vector, required if use_default is FALSE. Personalized parameters can be input here. Default is NULL.}

\item{Period}{numeric value. The time period for which the data needs to be estimated. Default is 24.}
}
\value{
a numeric vector that contains hyperparameters that can be used for \code{BayesDCirc}.
}
\description{
get_hyperparameters Function
}
\details{
We want to get hyperparameters using this function
}
\examples{
hyperparameters <- get_hyperparameters(Y1,Y2,t1,t2, use_default = T)
}
\author{
Yutao Zhang
}
