% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_main.R
\name{gibbs_main}
\alias{gibbs_main}
\title{gibbs sampling main function}
\usage{
gibbs_main(
  Y1,
  Y2,
  t1,
  t2,
  n.iter,
  hyperparameters,
  showIteration,
  priorInformation = NULL,
  priorFDR = NULL
)
}
\arguments{
\item{Y1}{expression values of group 1}

\item{Y2}{expression values of group 2}

\item{t1}{phenotype data of group 1}

\item{t2}{phenotype data of group 2}

\item{n.iter}{number of iteration}

\item{hyperparameters}{hyperparameters we need}

\item{showIteration}{logic value, if need show iteration when processing gibbs sampling}

\item{priorInformation}{a data frame with three columns. The first is gene's name, the second is the indicator for differential circadian gene with pattern difference, and the third is the indicator for differential circadian gene with MESOR difference. Default is NULL.}

\item{priorFDR}{The FDR cut off that prior information was used to claim differential circadian genes. Default is NULL.}
}
\value{
a list that contains posterior values
}
\description{
gibbs sampling
}
\details{
We want to use gibbs sampling to sample from posterior distributions for our parameter (Details)
}
\author{
Yutao Zhang
}
