% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top5per_delta.R
\name{top5per_delta}
\alias{top5per_delta}
\title{Get prior for delta}
\usage{
top5per_delta(mu_M)
}
\arguments{
\item{mu_M}{mu_M from linear regression}
}
\value{
prior
}
\description{
Get prior for delta
}
\details{
We want to get prior for delta
}
\examples{
mu_M = rnorm(1000)
top5per_delta(mu_M)
}
\author{
Yutao Zhang
}
