% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top5per_tau.R
\name{top5per_tau}
\alias{top5per_tau}
\title{Get prior for tau}
\usage{
top5per_tau(mu_E, mu_F)
}
\arguments{
\item{mu_E}{mu_E from linear regression}

\item{mu_F}{mu_F from linear regression}
}
\value{
prior
}
\description{
Get prior for tau
}
\details{
We want to get prior for tau
}
\examples{
mu_E = rnorm(1000)
mu_F = rnorm(1000)
top5per_tau(mu_E,mu_F)
}
\author{
Yutao Zhang
}
