% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFDR.R
\name{BayesianFDR}
\alias{BayesianFDR}
\title{compute the q values}
\usage{
BayesianFDR(belief)
}
\arguments{
\item{belief}{a vector, the ratio of sums of indicator for each gene to the iteration after burning}
}
\value{
q value
}
\description{
BayesianFDR Function
}
\details{
We want to compute the q values using this function
}
\examples{
belief = runif(1000)
BayesianFDR(belief)
}
\author{
Yutao Zhang
}
